requirements.txt
python-barcode
django-countries
Pillow
dateparser